tests: extract shared utility 'render_svg' method
