+++ tests/test_receipt.py
@@ -1,24 +1,9 @@
-import io
-
-from cairosvg import svg2png
 from PIL import Image
 from pyzbar.pyzbar import decode
 
 from apps.common.receipt import format_inline_qr, make_qr_png
 
-
-def render_svg(svg):
-    # pyzbar fails to decode qr codes under 52x52 and epc qr codes under 72x72
-    png = svg2png(bytestring=svg, parent_width=80, parent_height=80)
-    png_file = io.BytesIO(png)
-    image = Image.open(png_file)
-
-    # pillow renders alpha channel black by default
-    alpha = image.convert("RGBA").split()[-1]
-    opaque_image = Image.new("RGBA", image.size, "white")
-    opaque_image.paste(image, mask=alpha)
-
-    return opaque_image
+from tests._utils import render_svg
 
 
 def test_format_inline_qr():
